# openapi.yaml
openapi: 3.0.3
info:
  title: Karton API
  description: >
    API za aplikaciju "Karton" — vođenje auto servisa. Backend je jedini autoritet
    za poslovna pravila (specifikacija-finalna.md §7). Nazivi na engleskom (odluka 27),
    poruke grešaka na srpskom. Sva vremena u UTC; poslovna logika Europe/Belgrade.

    CSRF ugovor: svaki odgovor postavlja/osvežava XSRF-TOKEN cookie (čitljiv iz JS);
    svaki non-GET zahtev MORA poslati header X-CSRF-Token sa tom vrednošću —
    neslaganje ili izostanak → 403 CSRF_FAILED. Važi i za /auth/login (token se
    dobija prethodnim GET pozivom, npr. /auth/me).

    Optimističko zaključavanje: verzionisani agregati su work_order, document,
    appointment i settings. Sve mutacione akcije nad njima nose "version" u telu;
    zastarela verzija → 409 VERSION_CONFLICT. Akcije bez version polja (copy,
    DELETE termina, izdavanje predračuna) backend izvršava u transakciji uz
    row-lock i atomsku proveru aktuelnog stanja.
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /vehicles:
    get:
      tags: [vehicles]
      summary: Lista vozila; pretraga pogađa i stare tablice (BR-03)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
        - { name: status, in: query, schema: { type: string, enum: [active, archived] } }
        - { name: customerId, in: query, schema: { type: integer } }
        - { name: inShop, in: query, schema: { type: boolean }, description: Vozila sa aktivnim nalogom }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/VehiclePage' } } } }
    post:
      tags: [vehicles]
      summary: Novo vozilo (VIN duplikat 409 sa existingId; otvara istoriju vlasnika i tablice)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/VehicleInput' } } } }
      responses:
        '201': { description: Kreirano, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
        '409': { $ref: '#/components/responses/Duplicate' }
  /vehicles/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [vehicles]
      summary: Detalj vozila (trenutni vlasnik/tablica izvedeni iz aktivnih zapisa)
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } } }
    patch:
      tags: [vehicles]
      summary: Izmena (VIN se ne menja)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/VehicleInput' } } } }
      responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/Conflict' } }
  /vehicles/{id}/ownership:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [vehicles]
      responses: { '200': { description: Istorija vlasništva, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/OwnershipRecord' } } } } } }
    post:
      tags: [vehicles]
      summary: Promena vlasnika — zatvara aktivan zapis, otvara novi (transakciono)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerId, validFrom]
              properties:
                customerId: { type: integer }
                validFrom: { type: string, format: date }
      responses: { '201': { description: Promenjen vlasnik }, '422': { $ref: '#/components/responses/BusinessRule' } }
  /vehicles/{id}/registrations:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [vehicles]
      responses: { '200': { description: Istorija tablica, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/RegistrationRecord' } } } } } }
    post:
      tags: [vehicles]
      summary: Nova tablica — zatvara aktivan zapis, otvara novi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plate, validFrom]
              properties:
                plate: { type: string }
                validFrom: { type: string, format: date }
                note: { type: string }
      responses: { '201': { description: Evidentirano } }
  /vehicles/{id}/archive:
    post: { tags: [vehicles], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: Arhivirano } } }
  /vehicles/{id}/unarchive:
    post: { tags: [vehicles], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: Dearhivirano } } }

  # ---------- Mechanics ----------
...
  /vehicles/{id}/stats:
    get:
      tags: [vehicles]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      summary: Statistika vozila — broj naloga, ukupno naplaćeno (samo plaćeni računi, BR-31), poslednja poseta
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders: { type: integer }
                  totalSpent: { type: number }
                  lastVisit: { type: string, format: date, nullable: true }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: karton_session
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer } }
...
    Page: { name: page, in: query, schema: { type: integer, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, default: 20 }, description: Default iz Podešavanja }
    Sort: { name: sort, in: query, schema: { type: string }, description: 'npr. name:asc' }
    Query: { name: q, in: query, schema: { type: string }, description: Case-insensitive, ignoriše razmake }
  responses:
...
    Conflict:
      description: VERSION_CONFLICT — optimističko zaključavanje
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Duplicate:
      description: DUPLICATE_VIN / DUPLICATE_TAX_ID / DUPLICATE_EMAIL — nosi existingId
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    BusinessRule:
      description: 'Poslovno pravilo (422) — kod iz kataloga grešaka (spec §8)'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: PROFORMA_NOT_VALID }
        message: { type: string, example: 'Predračun je istekao i ne može se pretvoriti u račun.' }
        fields: { type: object, additionalProperties: { type: string } }
        warnings: { type: array, items: { type: string }, example: [MECHANIC_BUSY] }
        existingId: { type: integer, description: Kod duplikata — ID postojećeg entiteta }
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    Customer:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [individual, company] }
        name: { type: string }
        taxId: { type: string, nullable: true }
        address: { type: string }
        status: { type: string, enum: [active, archived] }
        contacts: { type: array, items: { $ref: '#/components/schemas/Contact' } }
...
    Contact:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, enum: [phone, email] }
        value: { type: string }
        isPrimary: { type: boolean }
...
    Vehicle:
      type: object
      properties:
        id: { type: integer }
        vin: { type: string }
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        fuel: { type: string }
        note: { type: string }
        status: { type: string, enum: [active, archived] }
        currentPlate: { type: string, description: Izvedeno iz aktivnog zapisa }
        currentOwner: { $ref: '#/components/schemas/Customer' }
    VehicleInput:
      type: object
      required: [vin, make, model]
      properties:
        vin: { type: string }
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        fuel: { type: string }
        plate: { type: string }
        ownerId: { type: integer }
        note: { type: string }
    VehiclePage:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Vehicle' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
    OwnershipRecord:
      type: object
      properties:
        id: { type: integer }
        customer: { $ref: '#/components/schemas/Customer' }
        validFrom: { type: string, format: date }
        validTo: { type: string, format: date, nullable: true }
    RegistrationRecord:
      type: object
      properties:
        id: { type: integer }
        plate: { type: string }
        validFrom: { type: string, format: date }
        validTo: { type: string, format: date, nullable: true }
        note: { type: string }
...